requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.3
attrs==21.4.0
boto3==1.35.6
botocore==1.35.6
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==2.2.0
Django==4.0.3
django-ipware==4.0.2
django-redis==5.4.0
django-storages==1.12.3
frozenlist==1.3.0
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
joblib==1.1.0
maxminddb==2.2.0
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.1
redis==5.0.8
requests==2.27.1
s3transfer==0.10.2
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==2.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
typing_extensions==4.12.2
urllib3==1.26.9
whitenoise==6.0.0
yarl==1.7.2

# Testing Dependencies
pytest==8.4.2
pytest-django==4.11.1
pytest-playwright==0.7.1
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
django-test-enforcer==0.1.0

# Utilities
rich==14.3.2
toml==0.10.2

